# packages/contracts/src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Agentj Control Plane API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://app.example.com"
    }
  ],
  "paths": {
...
    "/api/v1/projects/{projectId}/tunnels/{tunnelId}": {
      "delete": {
        "summary": "Stop and delete tunnel",
        "security": [{ "patAuth": [] }],
        "responses": { "204": { "description": "Tunnel deleted" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "patAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "PAT"
      }
    }
  }
}